feat: Hot reload python files when running locally
This PR does a couple of things.

First it reworks the Dockerfile for the API so that it runs expensive
commands first so we can take advantage of caching.

For instance it will install playwright and chromium up front since
those are slow and as it stands will always re-run whenever files
change.

Secondly it installs the python modules in a different folder then the
api code and sets the `PYTHONPATH` env variable to that folder so that
we can mount the `api` folder as a volume mounted to `/function/` in the
container.

Thirdly it installs `entr` and uses that to watch for changes to all
*.py files in the api folder and subdirectories, when those files change
it will restart aws-lambda-rie.

All of these changes taken together increase the dev-test loop for the
Api monolith from potentially five minutes down to almost
instantaneously.